APB-6728 [CS] move tax client count to tax group service
